refactor(database): split post data
